infra/cloud_functions/polygon_daily_loader/requirements.txt
# Cloud Function Dependencies for Polygon Daily Loader
# Python 3.11

# Google Cloud libraries
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.10.0
functions-framework==3.5.0

# Data processing
pandas==2.1.4
pyarrow==14.0.1

# HTTP requests
requests==2.31.0
